+++ src/test/java/org/catalytic/sdk/clients/AccessTokensTests.java
@@ -1,6 +1,7 @@
 package org.catalytic.sdk.clients;
 
 import org.catalytic.sdk.entities.AccessToken;
+import org.catalytic.sdk.entities.AccessTokensPage;
 import org.catalytic.sdk.exceptions.AccessTokenNotFoundException;
 import org.catalytic.sdk.exceptions.InternalErrorException;
 import org.catalytic.sdk.exceptions.UnauthorizedException;
@@ -13,6 +14,9 @@ import org.catalytic.sdk.generated.model.WaitForAccessTokenApprovalRequest;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import java.util.UUID;
 
 import static com.google.common.truth.Truth.assertThat;
@@ -20,11 +24,91 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 public class AccessTokensTests {
-    AccessTokensApi credentialsApi;
+    AccessTokensApi accessTokensApi;
 
     @Before
     public void before() {
-        credentialsApi = mock(AccessTokensApi.class);
+        accessTokensApi = mock(AccessTokensApi.class);
+    }
+
+    @Test(expected = AccessTokenNotFoundException.class)
+    public void list_itShouldReturnAccessTokenNotFoundExceptionIfClientInstantiatedWithoutToken() throws Exception {
+        AccessTokens accessTokensClient = new AccessTokens(null);
+        accessTokensClient.list();
+    }
+
+    @Test(expected = UnauthorizedException.class)
+    public void list_itShouldThrowUnauthorizedExceptionIfUnauthorized() throws Exception {
+        when(accessTokensApi.listAccessTokens(null, null)).thenThrow(new ApiException(401, null));
+
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
+        accessTokensClient.list();
+    }
+
+    @Test(expected = InternalErrorException.class)
+    public void list_itShouldThrowInternalErrorException() throws Exception {
+        when(accessTokensApi.listAccessTokens(null, null)).thenThrow(new ApiException(500, null));
+
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
+        accessTokensClient.list();
+    }
+
+    @Test
+    public void list_itShouldListAccessTokensWithOnePage() throws Exception {
+        org.catalytic.sdk.generated.model.AccessToken internalAccessToken = new org.catalytic.sdk.generated.model.AccessToken();
+        internalAccessToken.setId(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
+        internalAccessToken.setType(TokenType.USER);
+
+        org.catalytic.sdk.generated.model.AccessTokensPage internalAccessTokensPage = new org.catalytic.sdk.generated.model.AccessTokensPage();
+        List<org.catalytic.sdk.generated.model.AccessToken> accessTokens = Arrays.asList(internalAccessToken);
+
+        internalAccessTokensPage.setAccessTokens(accessTokens);
+        internalAccessTokensPage.setCount(accessTokens.size());
+        internalAccessTokensPage.setNextPageToken("");
+
+        when(accessTokensApi.listAccessTokens(null, null))
+                .thenReturn(internalAccessTokensPage);
+
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
+        AccessTokensPage accessTokensPage = accessTokensClient.list();
+        assertThat(accessTokensPage).isNotNull();
+        assertThat(accessTokensPage.getAccessTokens()).isNotEmpty();
+        assertThat(accessTokensPage.getAccessTokens().get(0).getId()).isEqualTo(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
+    }
+
+    @Test
+    public void list_itShouldListAccessTokensWithMoreThanOnePage() throws Exception {
+        org.catalytic.sdk.generated.model.AccessToken accessTokenOne = new org.catalytic.sdk.generated.model.AccessToken();
+        accessTokenOne.setId(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
+        accessTokenOne.setType(TokenType.USER);
+        List<org.catalytic.sdk.generated.model.AccessToken> accessTokenOneArray = Arrays.asList(accessTokenOne);
+
+        org.catalytic.sdk.generated.model.AccessToken accessTokenTwo = new org.catalytic.sdk.generated.model.AccessToken();
+        accessTokenTwo.setId(UUID.fromString("ac14952a-a331-457c-ac7d-9a284258b65a"));
+        accessTokenTwo.setType(TokenType.USER);
+        List<org.catalytic.sdk.generated.model.AccessToken> accessTokenTwoArray = Arrays.asList(accessTokenTwo);
+
+        org.catalytic.sdk.generated.model.AccessTokensPage pageOne = new org.catalytic.sdk.generated.model.AccessTokensPage();
+        pageOne.setAccessTokens(accessTokenOneArray);
+        pageOne.setCount(accessTokenOneArray.size());
+        pageOne.setNextPageToken("123abc");
+
+        org.catalytic.sdk.generated.model.AccessTokensPage pageTwo = new org.catalytic.sdk.generated.model.AccessTokensPage();
+        pageTwo.setAccessTokens(accessTokenTwoArray);
+        pageTwo.setCount(accessTokenTwoArray.size());
+        pageTwo.setNextPageToken("");
+
+        when(accessTokensApi.listAccessTokens(null, null))
+                .thenReturn(pageOne);
+
+        when(accessTokensApi.listAccessTokens("123abc", null))
+                .thenReturn(pageTwo);
+
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
+        AccessTokensPage accessTokensPage = accessTokensClient.list();
+        assertThat(accessTokensPage).isNotNull();
+        assertThat(accessTokensPage.getAccessTokens()).isNotEmpty();
+        assertThat(accessTokensPage.getCount()).isEqualTo(2);
     }
 
     @Test(expected = AccessTokenNotFoundException.class)
@@ -35,25 +119,25 @@ public class AccessTokensTests {
 
     @Test(expected = UnauthorizedException.class)
     public void getAccessToken_itShouldThrowUnauthorizedExceptionIfUnauthorized() throws Exception {
-        when(credentialsApi.getAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(401, null));
+        when(accessTokensApi.getAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(401, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         accessTokensClient.get("114c0d7d-c291-4ad2-a10d-68c5dd532af3");
     }
 
     @Test(expected = AccessTokenNotFoundException.class)
     public void getAccessToken_itShouldThrowUserNotFoundExceptionIfUserDoesNotExist() throws Exception {
-        when(credentialsApi.getAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(404, null));
+        when(accessTokensApi.getAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(404, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         accessTokensClient.get("114c0d7d-c291-4ad2-a10d-68c5dd532af3");
     }
 
     @Test(expected = InternalErrorException.class)
     public void getAccessToken_itShouldThrowInternalErrorException() throws Exception {
-        when(credentialsApi.getAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(500, null));
+        when(accessTokensApi.getAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(500, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         accessTokensClient.get("114c0d7d-c291-4ad2-a10d-68c5dd532af3");
     }
 
@@ -62,9 +146,9 @@ public class AccessTokensTests {
         org.catalytic.sdk.generated.model.AccessToken creds = new org.catalytic.sdk.generated.model.AccessToken();
         creds.setId(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
         creds.setType(TokenType.USER);
-        when(credentialsApi.getAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenReturn(creds);
+        when(accessTokensApi.getAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenReturn(creds);
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         AccessToken accessToken = accessTokensClient.get("114c0d7d-c291-4ad2-a10d-68c5dd532af3");
         assertThat(accessToken).isNotNull();
         assertThat(accessToken.getId()).isEqualTo(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
@@ -81,9 +165,9 @@ public class AccessTokensTests {
         org.catalytic.sdk.generated.model.AccessToken creds = new org.catalytic.sdk.generated.model.AccessToken();
         creds.setId(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
         creds.setType(TokenType.USER);
-        when(credentialsApi.revokeAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenReturn(creds);
+        when(accessTokensApi.revokeAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenReturn(creds);
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         AccessToken accessToken = accessTokensClient.revoke("114c0d7d-c291-4ad2-a10d-68c5dd532af3");
         assertThat(accessToken).isNotNull();
         assertThat(accessToken.getId()).isEqualTo(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
@@ -91,25 +175,25 @@ public class AccessTokensTests {
 
     @Test(expected = UnauthorizedException.class)
     public void revokeAccessTokens_itShouldThrowUnauthorizedExceptionIfUnauthorized() throws Exception {
-        when(credentialsApi.revokeAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(401, null));
+        when(accessTokensApi.revokeAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(401, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         accessTokensClient.revoke("114c0d7d-c291-4ad2-a10d-68c5dd532af3");
     }
 
     @Test(expected = AccessTokenNotFoundException.class)
     public void revokeAccessTokens_itShouldThrowUserNotFoundExceptionIfUserDoesNotExist() throws Exception {
-        when(credentialsApi.revokeAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(404, null));
+        when(accessTokensApi.revokeAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(404, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         accessTokensClient.revoke("114c0d7d-c291-4ad2-a10d-68c5dd532af3");
     }
 
     @Test(expected = InternalErrorException.class)
     public void revokeAccessTokens_itShouldThrowInternalErrorException() throws Exception {
-        when(credentialsApi.revokeAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(500, null));
+        when(accessTokensApi.revokeAccessToken("114c0d7d-c291-4ad2-a10d-68c5dd532af3")).thenThrow(new ApiException(500, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         accessTokensClient.revoke("114c0d7d-c291-4ad2-a10d-68c5dd532af3");
     }
 
@@ -119,9 +203,9 @@ public class AccessTokensTests {
         request.setDomain("example.pushbot.com");
         request.setEmail("alice@example.com");
         request.setPassword("mypassword");
-        when(credentialsApi.createAndApproveAccessToken(request)).thenThrow(new ApiException(401, null));
+        when(accessTokensApi.createAndApproveAccessToken(request)).thenThrow(new ApiException(401, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         accessTokensClient.create("example", "alice@example.com", "mypassword");
     }
 
@@ -131,9 +215,9 @@ public class AccessTokensTests {
         request.setDomain("example.pushbot.com");
         request.setEmail("alice@example.com");
         request.setPassword("mypassword");
-        when(credentialsApi.createAndApproveAccessToken(request)).thenThrow(new ApiException(500, null));
+        when(accessTokensApi.createAndApproveAccessToken(request)).thenThrow(new ApiException(500, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         accessTokensClient.create("example", "alice@example.com", "mypassword");
     }
 
@@ -146,9 +230,9 @@ public class AccessTokensTests {
         org.catalytic.sdk.generated.model.AccessToken creds = new org.catalytic.sdk.generated.model.AccessToken();
         creds.setId(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
         creds.setType(TokenType.USER);
-        when(credentialsApi.createAndApproveAccessToken(request)).thenReturn(creds);
+        when(accessTokensApi.createAndApproveAccessToken(request)).thenReturn(creds);
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         accessTokensClient.create("!@#$", "alice@example.com", "mypassword");
     }
 
@@ -161,9 +245,9 @@ public class AccessTokensTests {
         org.catalytic.sdk.generated.model.AccessToken creds = new org.catalytic.sdk.generated.model.AccessToken();
         creds.setId(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
         creds.setType(TokenType.USER);
-        when(credentialsApi.createAndApproveAccessToken(request)).thenReturn(creds);
+        when(accessTokensApi.createAndApproveAccessToken(request)).thenReturn(creds);
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         AccessToken accessToken = accessTokensClient.create("example", "alice@example.com", "mypassword");
         assertThat(accessToken).isNotNull();
         assertThat(accessToken.getId()).isEqualTo(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
@@ -178,9 +262,9 @@ public class AccessTokensTests {
         org.catalytic.sdk.generated.model.AccessToken creds = new org.catalytic.sdk.generated.model.AccessToken();
         creds.setId(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
         creds.setType(TokenType.USER);
-        when(credentialsApi.createAndApproveAccessToken(request)).thenReturn(creds);
+        when(accessTokensApi.createAndApproveAccessToken(request)).thenReturn(creds);
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         AccessToken accessToken = accessTokensClient.create("example.pushbot.com", "alice@example.com", "mypassword");
         assertThat(accessToken).isNotNull();
         assertThat(accessToken.getId()).isEqualTo(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
@@ -197,9 +281,9 @@ public class AccessTokensTests {
         creds.setId(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
         creds.setType(TokenType.USER);
         creds.setName("My AccessTokens");
-        when(credentialsApi.createAndApproveAccessToken(request)).thenReturn(creds);
+        when(accessTokensApi.createAndApproveAccessToken(request)).thenReturn(creds);
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         AccessToken accessToken = accessTokensClient.create("example", "alice@example.com", "mypassword", "My AccessTokens");
         assertThat(accessToken).isNotNull();
         assertThat(accessToken.getId()).isEqualTo(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
@@ -210,9 +294,9 @@ public class AccessTokensTests {
     public void createAccessTokensWithWebApprovalFlow_itShouldThrowUnauthorizedExceptionIfUnauthorized() throws Exception {
         AccessTokenCreationRequest request = new AccessTokenCreationRequest();
         request.setDomain("example.pushbot.com");
-        when(credentialsApi.createAccessToken(request)).thenThrow(new ApiException(401, null));
+        when(accessTokensApi.createAccessToken(request)).thenThrow(new ApiException(401, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         accessTokensClient.createWithWebApprovalFlow("example");
     }
 
@@ -220,9 +304,9 @@ public class AccessTokensTests {
     public void createAccessTokensWithWebApprovalFlow_itShouldThrowInternalErrorException() throws Exception {
         AccessTokenCreationRequest request = new AccessTokenCreationRequest();
         request.setDomain("example.pushbot.com");
-        when(credentialsApi.createAccessToken(request)).thenThrow(new ApiException(500, null));
+        when(accessTokensApi.createAccessToken(request)).thenThrow(new ApiException(500, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         accessTokensClient.createWithWebApprovalFlow("example");
     }
 
@@ -233,9 +317,9 @@ public class AccessTokensTests {
         org.catalytic.sdk.generated.model.AccessToken creds = new org.catalytic.sdk.generated.model.AccessToken();
         creds.setId(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
         creds.setType(TokenType.USER);
-        when(credentialsApi.createAccessToken(request)).thenReturn(creds);
+        when(accessTokensApi.createAccessToken(request)).thenReturn(creds);
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         AccessToken accessToken = accessTokensClient.createWithWebApprovalFlow("example");
         assertThat(accessToken).isNotNull();
         assertThat(accessToken.getId()).isEqualTo(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
@@ -250,9 +334,9 @@ public class AccessTokensTests {
         creds.setId(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
         creds.setType(TokenType.USER);
         creds.setName("My AccessTokens");
-        when(credentialsApi.createAccessToken(request)).thenReturn(creds);
+        when(accessTokensApi.createAccessToken(request)).thenReturn(creds);
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         AccessToken accessToken = accessTokensClient.createWithWebApprovalFlow("example", "My AccessTokens");
         assertThat(accessToken).isNotNull();
         assertThat(accessToken.getId()).isEqualTo(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"));
@@ -261,7 +345,7 @@ public class AccessTokensTests {
 
     @Test
     public void getApprovalUrl_itShouldReturnApprovalUrl() throws InternalErrorException {
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
 
         AccessToken accessToken = new AccessToken(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"), "example.com", null, null, null, null, null, null);
         String url = accessTokensClient.getApprovalUrl(accessToken);
@@ -270,7 +354,7 @@ public class AccessTokensTests {
 
     @Test
     public void getApprovalUrl_itShouldReturnApprovalUrlWithName() throws InternalErrorException {
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
 
         AccessToken accessToken = new AccessToken(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"), "example.com", null, "Example App", null, null, null, null);
         String url = accessTokensClient.getApprovalUrl(accessToken, "Example App");
@@ -282,9 +366,9 @@ public class AccessTokensTests {
         WaitForAccessTokenApprovalRequest request = new WaitForAccessTokenApprovalRequest();
         request.setToken("1234");
         request.setWaitTimeMillis(100);
-        when(credentialsApi.waitForAccessTokenApproval(request)).thenThrow(new ApiException(401, null));
+        when(accessTokensApi.waitForAccessTokenApproval(request)).thenThrow(new ApiException(401, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         AccessToken accessToken = new AccessToken(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"), "example", "user", "my token", "1234", "my-secret", "prod", "alice");
         accessTokensClient.waitForApproval(accessToken, 100);
     }
@@ -294,9 +378,9 @@ public class AccessTokensTests {
         WaitForAccessTokenApprovalRequest request = new WaitForAccessTokenApprovalRequest();
         request.setToken("1234");
         request.setWaitTimeMillis(100);
-        when(credentialsApi.waitForAccessTokenApproval(request)).thenThrow(new ApiException(500, null));
+        when(accessTokensApi.waitForAccessTokenApproval(request)).thenThrow(new ApiException(500, null));
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         AccessToken accessToken = new AccessToken(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"), "example", "user", "my token", "1234", "my-secret", "prod", "alice");
         accessTokensClient.waitForApproval(accessToken, 100);
     }
@@ -306,9 +390,9 @@ public class AccessTokensTests {
         WaitForAccessTokenApprovalRequest request = new WaitForAccessTokenApprovalRequest();
         request.setToken("1234");
         request.setWaitTimeMillis(100);
-        when(credentialsApi.waitForAccessTokenApproval(request)).thenReturn(null);
+        when(accessTokensApi.waitForAccessTokenApproval(request)).thenReturn(null);
 
-        AccessTokens accessTokensClient = new AccessTokens("1234", credentialsApi);
+        AccessTokens accessTokensClient = new AccessTokens("1234", accessTokensApi);
         AccessToken accessToken = new AccessToken(UUID.fromString("114c0d7d-c291-4ad2-a10d-68c5dd532af3"), "example", "user", "my token", "1234", "my-secret", "prod", "alice");
         accessTokensClient.waitForApproval(accessToken, 100);
     }